requirements-prod.txt
datasets>=2.19.0
pandas>=2.2.0
pyarrow>=15.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
requests>=2.31.0
